Game scene changes
Game scene has two temporary buttons which can be deleted later when win lose state implemented. They're just there for me to test game flow. Scene manager is now a singleton design pattern just for the sake that it's less lines of code to reference for the game state itself.